refactor: move and split command monitoring spec tests
